+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Moved properties on written shaders to a cbuffer. I don't think it does much, but it does make Unity complain less about SRP support
 - Added extra delay before updating the terrain maps because I just had the bug happen on me again for some reason
 - Some README and changelog notes updates
+- Tile size is now a Vector3 instead of a float
 
 ## v0.2.2
 - Made some clarifications in the changelog